requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.21.4
pillow==8.4.0
matplotlib==3.5.0
tqdm==4.62.3
scikit-learn==1.0.1
pandas==1.5.3
seaborn==0.12.0
json5==0.9.6 